--- backend/.python-version
@@ -1,1 +1,0 @@
-3.11.9
