+++ backend/app/routers/users.py
@@ -1,0 +1,99 @@
+from http import HTTPStatus
+
+from fastapi import APIRouter, Depends, HTTPException
+from sqlalchemy import select
+from sqlalchemy.orm import Session
+
+from app.database import get_session
+from app.models.user import User
+from app.schemas.request_body import UserSchema
+from app.schemas.response_body import Message, UserList, UserPublic
+from app.security import (
+    get_current_user,
+    get_password_hash,
+)
+
+router = APIRouter(prefix='/users', tags=['users'])
+
+
+@router.post('/', status_code=HTTPStatus.CREATED, response_model=UserPublic)
+def create_user(user: UserSchema, session: Session = Depends(get_session)):
+    db_user = session.scalar(
+        select(User).where(User.username == user.username)
+    )
+
+    if db_user:
+        raise HTTPException(
+            status_code=HTTPStatus.BAD_REQUEST,
+            detail='Username already registered',
+        )
+
+    hashed_password = get_password_hash(user.password)
+
+    db_user = User(
+        username=user.username, password=hashed_password, email=user.email
+    )
+    session.add(db_user)
+    session.commit()
+    session.refresh(db_user)
+
+    return db_user
+
+
+@router.get('/', status_code=HTTPStatus.OK, response_model=UserList)
+def read_users(
+    skip: int = 0, limit: int = 100, session: Session = Depends(get_session)
+):
+    users = session.scalars(select(User).offset(skip).limit(limit)).all()
+    return {'users': users}
+
+
+@router.put('/{user_id}', response_model=UserPublic)
+def update_user(
+    user_id: int,
+    user: UserSchema,
+    session: Session = Depends(get_session),
+    current_user: User = Depends(get_current_user),
+):
+    if current_user.id != user_id:
+        raise HTTPException(
+            status_code=HTTPStatus.BAD_REQUEST, detail='Not enough permissions'
+        )
+
+    db_user = session.scalar(select(User).where(User.id == user_id))
+    if not db_user:
+        raise HTTPException(
+            status_code=HTTPStatus.NOT_FOUND, detail='User not found'
+        )
+
+    db_user.username = user.username
+    db_user.password = get_password_hash(user.password)
+    db_user.email = user.email
+    session.commit()
+    session.refresh(db_user)
+
+    return db_user
+
+
+@router.delete('/{user_id}', response_model=Message)
+def delete_user(
+    user_id: int,
+    session: Session = Depends(get_session),
+    current_user: User = Depends(get_current_user),
+):
+    if current_user.id != user_id:
+        raise HTTPException(
+            status_code=HTTPStatus.BAD_REQUEST, detail='Not enough permissions'
+        )
+
+    db_user = session.scalar(select(User).where(User.id == user_id))
+
+    if not db_user:
+        raise HTTPException(
+            status_code=HTTPStatus.NOT_FOUND, detail='User not found'
+        )
+
+    session.delete(db_user)
+    session.commit()
+
+    return {'message': 'User deleted'}
